+++ .code/Sauce.cpp
@@ -1,2 +1,2 @@
 #include "Sauce.h"
-Sauce::Sauce(string nm, double p) :FoodItem(nm, p) {}
+Sauce::Sauce(string nm, float p) :FoodItem(nm, p) {}
