+++ .code/Sauce.h
@@ -1,10 +1,11 @@
 #ifndef SAUCE_H
 #define SAUCE_H
 #include "FoodItem.h"
+//first food item
 class Sauce :
     public FoodItem
 {
 public:
-    Sauce(string, double);
+    Sauce(string, float);
 };
 #endif
