+++ .code/Toppings.cpp
@@ -1,2 +1,2 @@
 #include "Toppings.h"
-Toppings::Toppings(string nm, double p) :FoodItem(nm, p) {}
+Toppings::Toppings(string nm, float p) :FoodItem(nm, p) {}
